Created Forms
I created three forms that will be needed for this project. The Login Form is for logging into the database. The Register Form is for registering as a new user for the program. The InvoiceDatabaseForm is for the database. It is unfinished as of now, but will be finished later.