Add files via upload
Updated the Relationship constraints, added code into the newWorkOrder Form, it should accept all of the users data and input it into the database. Still need to validate each field (aside from phone numbers), and still have to add the phone extension in. I think I forgot to do that as I think about it.