Updated RegisterForm, AdminForm, and WorkOrderForm
RegisterForm - Added Robbie's UI changes with slight tweaks to this project so everyone's work was in the same place. Also added code to the RegisterForm so everything displays as intended

AdminForm - Added a button to delete users, also added a connection to the ProjectDB along with two buttons that will  allow you to interact with the invoice table from that database. 

WorkOrderForm - Added the ability to print a copy of the selected invoice back in.